requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-cloud-speech==2.23.0
google-generativeai==0.3.2
openai-whisper==20231117
python-dotenv==1.0.0
pydub==0.25.1
werkzeug==3.0.1
soundfile==0.12.1
librosa==0.10.1
numpy==1.24.3
